feature:onboarding

+++ app/src/main/java/com/mit/learning_english/data/remote/dto/AudioResponse.kt
@@ -5,7 +5,7 @@ data class AudioResponse(
     val duration: Int,
     val format: String,
     val sampleRate: Int,
-    val fileSize: Int,
+    val fileSize: Double,
     val fileUrl: String,
     val pagesId: Int
 )
